# docs/swagger.yaml
definitions:
...
  controllers.DeleteResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/models.Photo'
      photoID:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.User:
    properties:
      age:
        type: integer
      ban_reason:
        type: string
      banned_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      last_seen_at:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      role:
        type: string
      social_medias:
        items:
          $ref: '#/definitions/models.SocialMedia'
        type: array
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/social-media:
    get:
      consumes:
      - application/json
      description: Get all social media in finalproject
      produces:
      - application/json
      responses:
        "200":
          description: Get all social media success
          schema:
            items:
              $ref: '#/definitions/models.SocialMedia'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all social media
      tags:
      - social media
    post:
      consumes:
      - application/json
      description: Create social media of the user
      parameters:
      - description: Social media request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Create social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create social media
      tags:
      - social media
  /api/v1/social-media/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Delete social media identified by given ID
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete social media success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid social media id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete social media
      tags:
      - social media
    get:
      consumes:
      - application/json
      description: Get social media identified by given id
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid social media id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get social media
      tags:
      - social media
    put:
      consumes:
      - application/json
      description: Update social media identified by given id
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      - description: Social media update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: Update social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update social media
      tags:
      - social media
...
  /socialmedia/create:
    post:
      consumes:
      - application/json
      description: Create social media of the user
      parameters:
      - description: Social media request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Create social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create social media
      tags:
      - social media
  /socialmedia/delete/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Delete social media identified by given ID
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete social media success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid social media id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete social media
      tags:
      - social media
  /socialmedia/get/{socialMediaId}:
    get:
      consumes:
      - application/json
      description: Get social media identified by given id
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid social media id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get social media
      tags:
      - social media
  /socialmedia/getall:
    get:
      consumes:
      - application/json
      description: Get all social media in finalproject
      produces:
      - application/json
      responses:
        "200":
          description: Get all social media success
          schema:
            items:
              $ref: '#/definitions/models.SocialMedia'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all social media
      tags:
      - social media
  /socialmedia/update/{socialMediaId}:
    put:
      consumes:
      - application/json
      description: Update social media identified by given id
      parameters:
      - description: ID of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      - description: Social media update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: Update social media success
          schema:
            $ref: '#/definitions/models.SocialMedia'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Social media not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update social media
      tags:
      - social media
...
swagger: "2.0"